regersoft/requirements.txt
aiogram==3.5.0
virtualbox
appium-python-client
requests
fastapi
uvicorn
pydantic